epic(core): de novo rebuild for python 3.14 native mandate
- Enforced python 3.14 via pyproject.toml, ruff targets, and mypy settings
- Flatted the `src/coreason_manifest/core/` directory as part of Epic 1
- Extracted foundational Shared Kernel types into `base.py` and `primitives.py`
- Removed execution-heavy dependencies while retaining `pydantic`
- Implemented pure type syntax based on PEP 695 in `primitives.py`
- Maintained a pristine, isolated `core/` scope awaiting next epic.
