+++ pyproject.toml
@@ -7,21 +7,17 @@ name = "coreason_manifest"
 version = "0.25.0"
 description = "This package is the definitive source of truth. If it isn't in the manifest, it doesn't exist. If it violates the manifest, it doesn't run."
 readme = "README.md"
-requires-python = ">=3.12"
+requires-python = ">=3.14"
 license = { file = "LICENSE" }
 authors = [
     { name = "Gowtham A Rao", email = "gowtham.rao@coreason.ai" },
 ]
 classifiers = [
     "License :: Other/Proprietary License",
-    "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.14",
     "Operating System :: OS Independent",
 ]
 dependencies = [
-    "fastmcp>=3.0.2",
-    "loguru>=0.7.3",
-    "mcp>=1.26.0",
-    "opentelemetry-api>=1.39.1",
     "pydantic>=2.12.5",
 ]
 
@@ -43,7 +39,6 @@ dev = [
     "pytest-asyncio>=1.3.0",
     "types-pyyaml>=6.0.12.20250915",
     "tzdata>=2025.3",
-    "atheris>=3.0.0",
     "pytest-benchmark>=5.2.3",
     "deepdiff>=8.6.1",
     "deptry>=0.23.0",
@@ -59,11 +54,11 @@ packages = ["src/coreason_manifest"]
 
 [tool.ruff]
 line-length = 120
-target-version = "py312"
+target-version = "py314"
 
 [tool.ruff.lint]
 select = [
-    "E", "F", "B", "I", "UP", "SIM", "RUF", "ARG", "C4", "PT", "TCH", "PIE", "RET",
+    "E", "F", "B", "I", "UP", "SIM", "RUF", "ARG", "C4", "PT", "TCH", "FA", "PIE", "RET",
     "PERF", "FURB", "LOG", "N", "A", "S"
 ]
 ignore = ["S101"]
@@ -72,7 +67,7 @@ ignore = ["S101"]
 known-first-party = ["coreason_manifest"]
 
 [tool.mypy]
-python_version = "3.12"
+python_version = "3.14"
 mypy_path = "src"
 strict = true
 disallow_untyped_defs = true
