--- src/coreason_manifest/core/primitives/constants.py
@@ -1,7 +1,0 @@
-from enum import StrEnum
-
-
-class NodeCapability(StrEnum):
-    COMPUTER_USE = "computer_use"
-    CODE_EXECUTION = "code_execution"
-    WASM_EXECUTION = "wasm_execution"
